+++ realm/realm-library/src/main/cpp/results_wrapper.hpp
