+++ realm/realm-library/src/objectServer/java/io/realm/SyncConfiguration.java
@@ -357,7 +357,7 @@ public class SyncConfiguration extends RealmConfiguration {
      * object you're interested in, a string containing a query determining which objects you want
      * to subscribe to, and a callback which will report the results.
      *
-     * @return {@true} is to open a partial synchronization Realm {@false} otherwise.
+     * @return {@code true} is to open a partial synchronization Realm {@code false} otherwise.
      */
     public boolean isPartialSync() {
         return isPartial;
