+++ realm/realm-library/src/syncIntegrationTest/java/io/realm/objectserver/model/PartialSyncModule.java
@@ -1,7 +1,23 @@
+/*
+ * Copyright 2017 Realm Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package io.realm.objectserver.model;
 
 import io.realm.annotations.RealmModule;
 
-@RealmModule(classes = { PartialSyncObjectA.class, PartialSyncObjectB.class})
+@RealmModule(classes = {PartialSyncObjectA.class, PartialSyncObjectB.class})
 public class PartialSyncModule {
 }
